👩‍💻 로그인

+++ src/main/java/com/example/tdw_backend/model/SignUpRequest.java
@@ -3,13 +3,11 @@ package com.example.tdw_backend.model;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
-
 @Getter
 @Setter
 public class SignUpRequest {
 
-    private BigInteger user_id;
+    private Long user_id;
 
     private String email;
